fix: 6 backend bugs — CUDA Windows tag, backend parsing, ctx precision, HF branch, file size, metric rename
- Windows CUDA backends: switch repo from ai-dock to ggml-org, keep b9279 default tag
- list_installed_backends: fix Windows backend parsing (win-cpu -> win, cpu)
- effective_ctx: use f64 to avoid precision loss for large context lengths
- list_gguf_files/fetch_readme: try main, fallback master on 404
- benchmark file_size_mb: use std::fs::metadata instead of placeholder 0
- rename first_token_time to prompt_processing_time (~27 refs)

+++ src/backend/server.rs
@@ -103,7 +103,7 @@ pub fn build_server_cmd(
         "--threads-batch",
         settings.threads_batch,
     );
-    let effective_ctx = (settings.context_length as f32 * settings.rope_scale) as u32;
+    let effective_ctx = (settings.context_length as f64 * settings.rope_scale as f64) as u32;
     push_arg(&mut cmd, &mut parts, "--ctx-size", effective_ctx);
     push_arg(&mut cmd, &mut parts, "--ubatch-size", settings.ubatch_size);
     if let Some(n) = settings.max_concurrent_predictions {
